Added confirmation for user deletion

+++ public/scripts/app.module.js
@@ -27,12 +27,15 @@ function MainController(NavService, MainService) {
 
   main.deleteUser = function() {
     console.log('Function works');
-    if(main.nav.state.loggedIn) {
-      console.log('Deleting user:', main.nav.userData.username);
-      MainService.deleteUser(main.nav.userData.username).then(function() {
-        console.log('Deleted user, logging out.');
-        main.logOutUser();
-      });
+    var confirmDelete = confirm('Are you sure you want to delete your account?');
+    if(confirmDelete) {
+      if(main.nav.state.loggedIn) {
+        console.log('Deleting user:', main.nav.userData.username);
+        MainService.deleteUser(main.nav.userData.username).then(function() {
+          console.log('Deleted user, logging out.');
+          main.logOutUser();
+        });
+      }
     }
   }
